move qemu stuff into a module
